# PB-962: Improve performance to list items